Added chunk content to be in the response to frontend, move weaviate logic to its own folder
